+++ src/shared/index.js
@@ -7,6 +7,7 @@ export { default as PillProp } from './types/PillProp';
 
 // functions
 export { default as makeCallback } from './functions/makeCallback';
+export { default as makeKeyboardCallback } from './functions/makeKeyboardCallback';
 export { default as makeIconFromProp } from './functions/makeIconFromProp';
 export { default as makeSpinnerFromProp } from './functions/makeSpinnerFromProp';
 export { default as makePillFromProp } from './functions/makePillFromProp';